+++ src/server/services/weather/weather-current.ts
@@ -15,11 +15,15 @@ export interface WeatherCurrent {
   dew_point: number;
   uvi: number;
   clouds: number;
-  rain?: number;
-  snow?: number;
+  rain?: {
+    '1h': number;
+  };
+  snow?: {
+    '1h': number;
+  };
   visibility: number;
   wind_speed: number;
   wind_deg: number;
   wind_gust: number;
-  weather: WeatherItem[];
+  weather: WeatherItem;
 }