+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Update OpenWeatherMap API from 2.5 to 3.0
+
 ## 3.0.0-alpha.92 (2024-10-13)
 
 * (theimo1221) Fix Temperature Sensor persistence